feat: (BREAKING) Create SG rule for client and rename variable
Since we don't want to have to handle the security groups rule for the
database on the client module. The easiest way is to give the
responsability of the destination to fill in the security group rules.

This adds that responsability on the rds module. Meaning that you no
longer have to create the security group rules on the client side.

The variables were renamed for better readability as well